+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next Release
 
 - Set a default icon for stellar objects, at least when changing their subtype
+- Allow rolling +hold from the preroll dialog if a move trigger is written that way
 
 ## 1.24.8